requirements.txt
torch=2.0.1
torchvision=0.15.2
numpy=1.24.3
Pillow=10.0.0
pycocotools=2.0.7
matplotlib=3.5.3
tensorboardX=2.6.2
tqdm
